Final Changes
Display port fro 4D systems display removed. All external peripherals
should be connected through the smart port and if level shifting is
required it must be included on the peripheral or if the peripheral is
off the shelf an adapter board needs to carry any level shifting or
additional circuitry.