Test Points and Labels
Test points added to each of the JTAG signals, as well as labels for
each test point.
Test points added for the erase signal and a lable
Test point added for reset as well as a label.
Test points for usb were implemented the same way as the mini-Rambo. DRC
check run and passed.