Fixes Implemented
Latest list of fixes from the last design review implemtned:
- New replacement crystal, high availability, same loading capacitance,
decent prize, 7M-12.000MEEQ-T
- MCU capacitors and decoupling rearranged and rerouted. In the Arduino
DUE implementation the filters are setup so the big cap and ferrite are
far from the MCU but the smaller cap is right up on the MCU, this
routing and placement scheme was adopted and implemented.
- Polygon pulled back on USB traces to reduce capacitive coupling and
noise
- USB placement and routing scheme copied from the Arduino DUE, all USB
traces are now on a single layer with no vias, minimizing impedance
changes between traces.